requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt

django==5.2.1
djangorestframework==3.16.0
openmeteo-requests==1.5.0
pandas==2.2.3
requests-cache==1.2.1
retry-requests==2.0.0
